+++ apps/desktop/electron/main.ts
@@ -3,7 +3,8 @@ import { createRequire } from 'node:module'
 import { fileURLToPath } from 'node:url'
 import path from 'node:path'
 import { ipcMain } from "electron";
-import { getCpuInfo, getMemoryInfo,getGpuInfo, getStorageInfo,getNetworkInfo,getStorageDevices } from "./services/system";
+import { getCpuInfo, getMemoryInfo,getGpuInfo, getStorageInfo,getNetworkInfo,getStorageDevices,getDisplayInfo } from "./services/system";
+import { getDevices } from "./services/devices";
 
 const require = createRequire(import.meta.url)
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
@@ -91,6 +92,14 @@ app.whenReady().then(() => {
   ipcMain.handle("system:network", async () => {
     return await getNetworkInfo();
   });
-  
+
+  ipcMain.handle("system:getDisplayInfo", async () => {
+    return getDisplayInfo();
+  });
+  ipcMain.handle("system:getDevices", async () => {
+    return getDevices();
+  });
+
+
   createWindow();
 });